pyproject.toml
[tool.poetry]
name = "aletheia"
version = "0.1.0"
description = ""
authors = ["Joshua \"jag\" Ginsberg <jag@flowtheory.net>"]

[tool.poetry.dependencies]
python = "~3.6.1"
semver = "^2.10.0"
beautifulsoup4 = "^4.9.0"
pyyaml = "^5.3.1"
google-api-python-client = "^1.8.3"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
python-dateutil = "^2.8.1"
setuptools = ">=40.3.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.1"
pre-commit = "^2.4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
